Adjusted width so that the toolbar should not be wider than the parent. Removes the 4px border we added to the parent. [Actually remeasures the border on the parent to allow for more flexible stylesheet.  Only called once so no performance hit.]
Note: No precautions are teken (yet) to deal with padding inside the element, which would break a few things.

My attempts to use content-box to get the accurate inner-width did not work,
Mootools includes the border and pading when getting the coordinates.
Attempts to use content-box and setting the el to 100% also failed, due to the funny way browsers calculate 100%.]


Signed-off-by: samgoody <ssgoodman@gmail.com>